Elimină PaymentScreenStatus.js și înlocuiește cu versiunea ESM.
Scriptul `PaymentScreenStatus.js` a fost eliminat și înlocuit cu un fișier ESM pentru a respecta structura modernizată a codului POS. Această schimbare include ajustări în logică, migrare către un sistem bazat pe patch-uri și actualizări aferente în manifest și fișierul XML corelat.
