+++ l10n_ro_pos/static/src/xml/pos.xml
@@ -5,7 +5,7 @@
         t-inherit="point_of_sale.PaymentScreenStatus"
         t-inherit-mode="extension"
     >
-        <xpath expr="//div[hasclass('payment-status-container')]" position="before">
+        <xpath expr="//section" position="before">
             <div t-if="warningMessage=='L5000'" class="limit-warning">
                 <span>Sale to individuals only..</span>
             </div>
